% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev_estimation.R
\name{bgev_log_likelihood}
\alias{bgev_log_likelihood}
\title{Log-likelihood function for the BGEV distribution}
\usage{
bgev_log_likelihood(x, pars)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{pars}{Vector of parameters (mu, sigma, xi, delta). See \link{bgev}.}
}
\value{
The log-likelihood value.
}
\description{
Log-likelihood function for the BGEV distribution
}
\author{
Thiago do Rego Sousa and Yasmin Lirio
}
