% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev_domain.R
\name{bgev_support}
\alias{bgev_support}
\title{Compute the support of the BGEV distribution}
\usage{
bgev_support(mu = 1, sigma = 1, xi = 0.3, delta = 2)
}
\arguments{
\item{mu}{location parameter}

\item{sigma}{scale parameter (sigma > 0)}

\item{xi}{shape parameter in R}

\item{delta}{shape parameter (delta > -1)}
}
\value{
A vector of length 2 with the lower and upper limits of the support
}
\description{
Returns the lower and upper limits of the support of the BGEV distribution
}
\details{
It returns values with \code{-Inf} or \code{Inf} when the support is unbounded. 
When the shape parameter \code{xi} is different from zero, the support 
is truncated either at the left or at the right side of the real. 
Considering the support is particularly useful to estimating moments and
to compute the likelihood function.
}
\author{
Thiago do Rego Sousa
}
