% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev_domain.R
\name{bgev_valid_params}
\alias{bgev_valid_params}
\title{Validate BGEV parameters}
\usage{
bgev_valid_params(mu, sigma, xi, delta)
}
\arguments{
\item{mu}{location parameter, as in \link{bgev}}

\item{sigma}{scale parameter (sigma > 0), as in \link{bgev}}

\item{xi}{shape parameter in R, as in \link{bgev}}

\item{delta}{shape parameter (delta > -1), as in \link{bgev}}
}
\description{
Check if the provided parameters for the BGEV distribution are valid.
}
\author{
Thiago do Rego Sousa
}
