% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgev.R
\name{dbgev}
\alias{dbgev}
\title{Bimodal Generalized Extreme Value: Density Function}
\usage{
dbgev(x, csi, mu, sigma, delta)
}
\arguments{
\item{x}{Domain, x in R.}

\item{csi}{Shape parameter, csi != 0.}

\item{mu}{Location parameter, mu in R.}

\item{sigma}{Scale parameter, sigma > 0.}

\item{delta}{Second location parameter, delta >= -1.}
}
\value{
Vector.
}
\description{
Bimodal Generalized Extreme Value: Density Function
}
\examples{
dbgev(x = -2, csi = -1, mu = 1, sigma = 1, delta = 1)
curve(dbgev(x, csi = -1, mu = 1, sigma = 1, delta = 1), xlim = c(-3, 2))
integrate(dbgev, csi = -1, mu = 1, sigma = 1, delta = 1, lower = -Inf, upper = Inf)
}
