% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgev.R
\name{dgev}
\alias{dgev}
\title{Generalized Extreme Value: Density Function}
\usage{
dgev(x, csi, mu, sigma)
}
\arguments{
\item{x}{Domain.}

\item{csi}{Shape parameter.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\value{
Vector.
}
\description{
Generalized Extreme Value: Density Function
}
\examples{
dgev(x = 2, csi = 0, mu = 1, sigma = 1)
curve(dgev(x, csi = 0, mu = 1, sigma = 1), xlim = c(-5, 10), ylim= c(0, .4))
integrate(dgev, csi = 0, mu = 1, sigma = 1, lower = -5, upper = 0)
}
