% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgev.R
\name{pgev}
\alias{pgev}
\title{Generalized Extreme Value: Distribution Function}
\usage{
pgev(q, csi, mu, sigma, lower.tail = TRUE)
}
\arguments{
\item{q}{Quantile.}

\item{csi}{Shape parameter.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X <= x) otherwise, P(X > x).}
}
\value{
Vector.
}
\description{
Generalized Extreme Value: Distribution Function
}
\examples{
pgev(0, csi = 0, mu = 1, sigma = 1)
integrate(dbgev, csi = 0, mu = 1, sigma = 1, lower = -Inf, upper = 0)
pgev(0, csi = 0, mu = 1, sigma = 1, lower.tail = FALSE)
curve(pgev(x, csi = 0, mu = 1, sigma = 1), xlim = c(-5, 10))
}
