% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T0.R
\name{T0}
\alias{T0}
\title{Generalized Extreme Value: T0}
\usage{
T0(x, sigma, delta)
}
\arguments{
\item{x}{Domain, x in R.}

\item{sigma}{Scale parameter, sigma > 0.}

\item{delta}{Second location parameter, delta >= -1.}
}
\value{
Vector.
}
\description{
Generalized Extreme Value: T0
}
\examples{
T0(x = 2, sigma = 1, delta = 1)
}
