% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T1.R
\name{T1}
\alias{T1}
\title{Generalized Extreme Value: T1. First derivative of T0}
\usage{
T1(x, sigma, delta)
}
\arguments{
\item{x}{Domain, x in R.}

\item{sigma}{Scale parameter, sigma > 0.}

\item{delta}{Second location parameter, delta >= -1.}
}
\value{
Vector.
}
\description{
Generalized Extreme Value: T1. First derivative of T0
}
\examples{
T1(x = 2, sigma = 1, delta = 1)
}
