% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2.R
\name{T2}
\alias{T2}
\title{Generalized Extreme Value: T2. Second derivative of T0}
\usage{
T2(x, sigma, delta)
}
\arguments{
\item{x}{Domain, x in R.}

\item{sigma}{Scale parameter, sigma > 0.}

\item{delta}{Second location parameter, delta >= -1.}
}
\value{
Vector.
}
\description{
Generalized Extreme Value: T2. Second derivative of T0
}
\examples{
T2(x = 2, sigma = 1, delta = 1)
}
