% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tinv.R
\name{Tinv}
\alias{Tinv}
\title{Generalized Extreme Value: Tinv. Inverse function of T0}
\usage{
Tinv(x, sigma, delta)
}
\arguments{
\item{x}{Domain, x in R.}

\item{sigma}{Scale parameter, sigma > 0.}

\item{delta}{Second location parameter, delta >= -1.}
}
\value{
Vector.
}
\description{
Generalized Extreme Value: Tinv. Inverse function of T0
}
\examples{
Tinv(x = 2, sigma = 1, delta = 1)
}
