% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev_distribution.R
\name{bgev}
\alias{bgev}
\alias{dbgev}
\alias{pbgev}
\alias{qbgev}
\alias{rbgev}
\title{Bimodal GEV (generalized extreme value) distribution}
\usage{
dbgev(x, mu = 1, sigma = 1, xi = 0.3, delta = 2)
}
\arguments{
\item{x}{Numeric vector of values for calculating density.}

\item{mu}{location parameter}

\item{sigma}{scale parameter (sigma > 0)}

\item{xi}{shape parameter in R}

\item{delta}{shape parameter (delta > -1)}

\item{q}{Numeric vector of quantiles.}

\item{p}{Numeric vector of probabilities.}

\item{n}{Number of observations for random generation.}

\item{log}{Logical; if \code{TRUE}, densities are returned on the log scale.}
}
\value{
\item{dbgev}{density values}
\item{pbgev}{distribution function values}
\item{qbgev}{quantile function values}
\item{rbgev}{random variates}
}
\description{
Functions to compute the density, distribution function, quantile function,
and to generate random variates for the bgebv (bimodal generalized extreme value)
}
\details{
This distribution corresponds was proposed by in Cira EG Otiniano, 
Bianca S Paiva, Roberto Vila and Marcelo Bourguignon (2021)
}
\note{
BGEV distribution is equivalent to the GEV distribution when \code{delta = 0}. 
When comparing BGEV with GEV from package EnvStats, the \code{shape} parameter of GEV
is changed to \code{-xi} due to reparametrization
}
\examples{
par(mfrow = c(2, 2))
set.seed(1000)
r <- rbgev(n = 1000)
plot(r, type = "l", main = "BGEV Random Values")

hist(r, probability = TRUE, border = "white", ylim = c(0,1))
x <- seq(min(r), max(r), length = 201)
lines(x, dbgev(x), lwd = 2)

plot(sort(r), (1:1000)/1000, main = "Probability", ylab = "Probability")
lines(x, pbgev(x), lwd = 2)

round(qbgev(pbgev(q = seq(0, 3, by = 0.1)), 6),2)
}
\references{
Otiniano, Cira E. G., et al. (2023).
\emph{A bimodal model for extremes data}.
Environmental and Ecological Statistics, 1–28.
\doi{10.1007/s10651-023-00566-7}
}
\author{
Thiago do Rego Sousa and Yasmin Lirio
}
