% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev_start_params.R
\name{bgev_start_using_quantiles}
\alias{bgev_start_using_quantiles}
\title{Starting values for BGEV distribution}
\usage{
bgev_start_using_quantiles(x)
}
\arguments{
\item{x}{Numeric vector of observations.}
}
\value{
starts_DEoptim start values for the estimation of (mu,sigma,xi,delta)
}
\description{
Starting values for BGEV distribution
}
\author{
Thiago do Rego Sousa
}
