% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev_domain.R
\name{bgev_valid_params}
\alias{bgev_valid_params}
\title{Validate BGEV parameters}
\usage{
bgev_valid_params(mu, sigma, xi, delta)
}
\arguments{
\item{\link{bgev}}{Parameters of the BGEV distribution as in \link{bgev}}
}
\description{
Check if the provided parameters for the BGEV distribution are valid.
}
\author{
Thiago do Rego Sousa
}
