% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgev.R
\name{dbgev}
\alias{dbgev}
\title{Bimodal Generalized Extreme Value: Density Function}
\usage{
dbgev(x, csi, mu, sigma, delta)
}
\arguments{
\item{x}{Domain.}

\item{csi}{Shape parameter.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{delta}{Second location parameter.}
}
\value{
Vector.
}
\description{
Bimodal Generalized Extreme Value: Density Function
}
\examples{
dbgev(x = 2, csi = 0, mu = 1, sigma = 1, delta = 1)
curve(dbgev(x, csi = 0, mu = 1, sigma = 1, delta = 1), xlim = c(-5, 10))
integrate(dbgev, csi = 0, mu = 1, sigma = 1, delta = 1, lower = -5, upper = 0)
dbgev(x = 2, csi = 1, mu = 0, sigma = 1, delta = 2)
curve(dbgev(x, csi = 1, mu = 0, sigma = 1, delta = 2), xlim = c(-1, 5))
}
