% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbgev.R
\name{pbgev}
\alias{pbgev}
\title{Bimodal Generalized Extreme Value: Distribution Function}
\usage{
pbgev(q, csi, mu, sigma, delta, lower.tail = TRUE)
}
\arguments{
\item{q}{Quantile.}

\item{csi}{Shape parameter.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{delta}{Second location parameter.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X <= x) otherwise, P(X > x).}
}
\value{
Vector.
}
\description{
Bimodal Generalized Extreme Value: Distribution Function
}
\examples{
pbgev(0, csi = 1, mu = 2, sigma = 1, delta = 2)
pbgev(0, csi = 1, mu = 2, sigma = 1, delta = 2, lower.tail = FALSE)
curve(pbgev(x, csi = 1, mu = 2, sigma = 1, delta = 2), xlim = c(-5, 10))
}
